update of the search component
