+++ eldorado-pokedex/src/pages/Pokédex/Cards/Cards.js
@@ -1,5 +1,4 @@
-import React, { useEffect, useState } from "react";
-import { getAllPokemon, getPokemon } from "../../../service/api";
+import React from "react";
 
 import {
   ContainerCard,
@@ -13,56 +12,7 @@ import {
 } from "./StylesCard";
 import { Container, Item } from "../Pagination/StylesPagination";
 
-export default function Cards() {
-  const [pokemonData, setPokemonData] = useState([]);
-  const [active, setActive] = useState(1);
-  const [currentOffset, setCurrentOffset] = useState(0);
-  const offset = 9;
-  const initialURL = `https://pokeapi.co/api/v2/pokemon?limit=9S&offset=${
-    currentOffset * offset
-  }`;
-
-  useEffect(() => {
-    async function fetchData() {
-      let response = await getAllPokemon(initialURL);
-      await loadPokemon(response.results);
-    }
-    fetchData();
-  }, []);
-
-  const loadPokemon = async (data) => {
-    let _pokemonData = await Promise.all(
-      data.map(async (pokemon) => {
-        let pokemonRecord = await getPokemon(pokemon);
-        return pokemonRecord;
-      })
-    );
-    setPokemonData(_pokemonData);
-  };
-
-  const nextPagePokemon = async () => {
-    const currentOffsetAux = currentOffset + 1;
-    setCurrentOffset(currentOffset + 1);
-    setActive(active + 1);
-    let response = await getAllPokemon(
-      `https://pokeapi.co/api/v2/pokemon?limit=9S&offset=${
-        currentOffsetAux * offset
-      }`
-    );
-    await loadPokemon(response.results);
-  };
-
-  const previousPagePokemon = async () => {
-    const currentOffsetAux = currentOffset - 1;
-    setCurrentOffset(currentOffset - 1);
-    setActive(active - 1);
-    let response = await getAllPokemon(
-      `https://pokeapi.co/api/v2/pokemon?limit=9S&offset=${
-        currentOffsetAux * offset
-      }`
-    );
-    await loadPokemon(response.results);
-  };
+export default function Cards({pokemonData ,nextPagePokemon,previousPagePokemon,active}) {
 
   return (
     <>
