+++ eldorado-pokedex/src/pages/Pokédex/Search/enjoy.js
@@ -1,0 +1,11 @@
+ //   const handleSearchPokemon = () =>{
+  //      const pokemonFilter = pokemonData.filter(pokemons=>pokemons.name===search)
+
+  //      setPokemonData(pokemonFilter)
+  // }
+
+  useEffect(() => {
+    if (search==="") {
+      setPokemonData(POKEMON_DEFAULT);
+    }
+  },[search]);
