+++ eldorado-pokedex/src/service/api.js
@@ -1,5 +1,5 @@
 
-  export function getPokemon({ url }) {
+  export function getPokemon( {url} ) {
     return new Promise((resolve, reject) => {
         fetch(url).then(res => res.json())
             .then(data => {
